refactor(renderer): organize document modules
