+++ apps/desktop/src/renderer/documents/artifacts/PdfPreview.tsx
