+++ apps/desktop/src/renderer/documents/editable/editor/DocumentStatusBar.tsx
