+++ apps/desktop/src/renderer/docx-worker.html
@@ -6,5 +6,5 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>JobOS DOCX Worker</title>
 </head>
-<body><div id="worker-root" aria-hidden="true"></div><script type="module" src="/docxWorker.ts"></script></body>
+<body><div id="worker-root" aria-hidden="true"></div><script type="module" src="/documents/docx/worker/docxWorker.ts"></script></body>
 </html>
