+++ apps/desktop/src/renderer/print.html
@@ -9,6 +9,6 @@
   <body>
     <div id="print-source" hidden></div>
     <div id="print-pages"></div>
-    <script type="module" src="/documentPrint.ts"></script>
+    <script type="module" src="/documents/editable/print/documentPrint.ts"></script>
   </body>
 </html>
